+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M346Z168A4VXE28FGX3KWA77"),
+  id: z.string().default("01M36FZEAF5WE29P0HE6E4F8VY"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>