+++ packages/orq-rc/RELEASES.md
@@ -12569,3 +12569,13 @@ Based on:
 - [typescript v4.16.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.12 - packages/orq-rc
+
+## 2026-09-23 06:44:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.798.0 (2.938.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.13 - packages/orq-rc